Optimizations for large examples
The key point is to look for failures near the beginning of the string
and to only lazily build as much of the history as we need. This means
that when we're faced with a very large example we are mostly not forced
to process much of it, and also decreases the number of binary search
steps we have to perform.